refactor(web): split chat panels under the file-size cap
The File size check CI gate flagged three files pushed over the 500-line
review cap by PR C:
  - web/src/panels/Chat.svelte (609)
  - web/src/panels/Chat.test.js (638)
  - web/src/panels/ChannelTimeline.svelte (527)

Split each along a coherent seam rather than grandfathering (matching the
repo's split-don't-grow convention):
  - Extract the session-scope selector (§C) into ScopeSelector.svelte, a
    child component bound back to the composer via sessionId/epochId.
  - Hoist the duplicated formatTimestamp plus channelLabel/senderLabel into
    a shared lib/format.js (with a unit test), used by both panels.
  - Move the composer + session-scope tests into Chat.composer.test.js.

No behavior change; web build + all 112 tests pass.
